+++ src/Pages/User/Task/TaskList.js
@@ -3,6 +3,7 @@ import React, { useEffect } from 'react';
 import { Button, Row } from 'react-bootstrap';
 import { useDispatch, useSelector } from 'react-redux';
 import { NavLink, useNavigate } from 'react-router-dom';
+import swal from 'sweetalert';
 import TaskCard from '../../../components/TaskCard/TaskCard';
 import urlString from '../../../constant/urlString';
 import { setTask } from '../../../redux/action/taskAction';
@@ -20,10 +21,22 @@ function Task() {
 
     //delete task
     const handelDelete = (id) => {
-        axios.delete('http://localhost:5050/task/' + id).then(res => {
-            console.log(res)
-            getAlltask()
+        swal({
+            title: "Are you sure?",
+            text: "you want to delete this",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
         })
+            .then((willDelete) => {
+                if (willDelete) {
+                    axios.delete('http://localhost:5050/task/' + id).then(res => {
+                        console.log(res)
+                        getAlltask()
+                    })
+                }
+            });
+
     }
 
     //edit task
@@ -52,9 +65,8 @@ function Task() {
             <div>
                 <NavLink to={urlString.TASKCREATE}><Button variant="success my-3">Add</Button></NavLink>
             </div>
-            <Row className='d-flex justify-content-center'>
+            <Row>
                 {tasks && tasks.map((data, index) =>
-
                     <TaskCard data={data} handelDelete={handelDelete} handelEdit={handelTaskEdit} key={index} />
                 )}
             </Row>
